Moved Navigation into components folder
